+++ peer/roundrobin/peerring.go
@@ -26,23 +26,23 @@ import (
 	"go.uber.org/yarpc/api/peer"
 )
 
-// NewPeerRing creates a new PeerRing with an initial capacity
-func NewPeerRing(capacity int) *PeerRing {
-	return &PeerRing{
+// newPeerRing creates a new peerRing with an initial capacity
+func newPeerRing(capacity int) *peerRing {
+	return &peerRing{
 		peerToNode: make(map[string]*ring.Ring, capacity),
 	}
 }
 
-// PeerRing provides a safe way to interact (Add/Remove/Get) with a potentially
+// peerRing provides a safe way to interact (Add/Remove/Get) with a potentially
 // changing list of peer objects
-// PeerRing is NOT Thread-safe, make sure to only call PeerRing functions with a lock
-type PeerRing struct {
+// peerRing is NOT Thread-safe, make sure to only call peerRing functions with a lock
+type peerRing struct {
 	peerToNode map[string]*ring.Ring
 	nextNode   *ring.Ring
 }
 
 // GetPeer returns the Peer from the Ring or Nil
-func (pr *PeerRing) GetPeer(pid peer.Identifier) peer.Peer {
+func (pr *peerRing) GetPeer(pid peer.Identifier) peer.Peer {
 	node, ok := pr.peerToNode[pid.Identifier()]
 	if !ok {
 		return nil
@@ -51,9 +51,9 @@ func (pr *PeerRing) GetPeer(pid peer.Identifier) peer.Peer {
 	return getPeerForRingNode(node)
 }
 
-// Add a peer.Peer to the end of the PeerRing, if the ring is empty
+// Add a peer.Peer to the end of the peerRing, if the ring is empty
 // it initializes the nextNode marker
-func (pr *PeerRing) Add(p peer.Peer) error {
+func (pr *peerRing) Add(p peer.Peer) error {
 	if _, ok := pr.peerToNode[p.Identifier()]; ok {
 		// Peer Already in ring, ignore the add
 		return peer.ErrPeerAddAlreadyInList(p.Identifier())
@@ -78,9 +78,9 @@ func newPeerRingNode(p peer.Peer) *ring.Ring {
 	return newNode
 }
 
-// Remove a peer Peer from the PeerRing, if the PeerID is not
+// Remove a peer Peer from the peerRing, if the PeerID is not
 // in the ring return an error
-func (pr *PeerRing) Remove(p peer.Peer) error {
+func (pr *peerRing) Remove(p peer.Peer) error {
 	node, ok := pr.peerToNode[p.Identifier()]
 	if !ok {
 		// Peer doesn't exist in the list
@@ -93,7 +93,7 @@ func (pr *PeerRing) Remove(p peer.Peer) error {
 }
 
 // RemoveAll pops all the peers from the ring and returns them in a list
-func (pr *PeerRing) RemoveAll() []peer.Peer {
+func (pr *peerRing) RemoveAll() []peer.Peer {
 	peers := make([]peer.Peer, 0, len(pr.peerToNode))
 	for _, node := range pr.peerToNode {
 		peers = append(peers, pr.popNode(node))
@@ -102,7 +102,7 @@ func (pr *PeerRing) RemoveAll() []peer.Peer {
 }
 
 // All returns a snapshot of all the peers from the ring as a list.
-func (pr *PeerRing) All() []peer.Peer {
+func (pr *peerRing) All() []peer.Peer {
 	peers := make([]peer.Peer, 0, len(pr.peerToNode))
 	for _, node := range pr.peerToNode {
 		peers = append(peers, getPeerForRingNode(node))
@@ -110,7 +110,7 @@ func (pr *PeerRing) All() []peer.Peer {
 	return peers
 }
 
-func (pr *PeerRing) popNode(node *ring.Ring) peer.Peer {
+func (pr *peerRing) popNode(node *ring.Ring) peer.Peer {
 	p := getPeerForRingNode(node)
 
 	if isLastRingNode(node) {
@@ -128,13 +128,13 @@ func (pr *PeerRing) popNode(node *ring.Ring) peer.Peer {
 	return p
 }
 
-func (pr *PeerRing) isNextNode(node *ring.Ring) bool {
+func (pr *peerRing) isNextNode(node *ring.Ring) bool {
 	return pr.nextNode == node
 }
 
 // Next returns the next peer in the ring, or nil if there is no peer in the ring
 // after it has the next peer, it increments the nextPeer marker in the ring
-func (pr *PeerRing) Next() peer.Peer {
+func (pr *peerRing) Next() peer.Peer {
 	if pr.nextNode == nil {
 		return nil
 	}